=str #16751 Update to reactive-streams 1.0-RC3
Changed rules:
* 1.9 Always onSubscribe prior to any other signals
* 1.9 NullPointerException if subscriber is null
* 3.17 Long overflow, effectively unbounded instead of onError

Fixed some more things:
* fixed some FIXME
* Long drop and take
* memory leaks in tck tests, use BeforeClass to create ActorSystem
  use AfterClass to shutdown ActorSystem
* more tck tests
* don't emit OnComplete when substream is cancelled
* work around for memory leak in PrefixAndTail

+++ akka-stream/src/main/scala/akka/stream/impl/fusing/Ops.scala
@@ -75,8 +75,8 @@ private[akka] final case class MapConcat[In, Out](f: In ⇒ immutable.Seq[Out], 
 /**
  * INTERNAL API
  */
-private[akka] final case class Take[T](count: Int) extends PushStage[T, T] {
-  private var left: Int = count
+private[akka] final case class Take[T](count: Long) extends PushStage[T, T] {
+  private var left: Long = count
 
   override def onPush(elem: T, ctx: Context[T]): Directive = {
     left -= 1
@@ -89,8 +89,8 @@ private[akka] final case class Take[T](count: Int) extends PushStage[T, T] {
 /**
  * INTERNAL API
  */
-private[akka] final case class Drop[T](count: Int) extends PushStage[T, T] {
-  private var left: Int = count
+private[akka] final case class Drop[T](count: Long) extends PushStage[T, T] {
+  private var left: Long = count
   override def onPush(elem: T, ctx: Context[T]): Directive =
     if (left > 0) {
       left -= 1
